# improve flag parsing performance