--- a/pkg/events/parse_args_test.go
+++ b/pkg/events/parse_args_test.go
@@ -13,6 +13,78 @@ import (
 func TestParseArgs(t *testing.T) {
 	t.Parallel()
 
+	t.Run("Parse pointer value", func(t *testing.T) {
+		t.Parallel()
+
+		testCases := []struct {
+			name         string
+			args         []trace.Argument
+			expectedArgs []trace.Argument
+		}{
+			{
+				name: "ptrace addr arg",
+				args: []trace.Argument{
+					{
+						ArgMeta: trace.ArgMeta{
+							Name: "addr",
+							Type: "void*",
+						},
+						Value: ^uintptr(0),
+					},
+				},
+				expectedArgs: []trace.Argument{
+					{
+						ArgMeta: trace.ArgMeta{
+							Name: "addr",
+							Type: "void*",
+						},
+						Value: "0xffffffffffffffff",
+					},
+				},
+			},
+			{
+				name: "ptrace addr arg",
+				args: []trace.Argument{
+					{
+						ArgMeta: trace.ArgMeta{
+							Name: "addr",
+							Type: "void*",
+						},
+						Value: uintptr(0x42424242),
+					},
+				},
+				expectedArgs: []trace.Argument{
+					{
+						ArgMeta: trace.ArgMeta{
+							Name: "addr",
+							Type: "void*",
+						},
+						Value: "0x42424242",
+					},
+				},
+			},
+		}
+
+		for _, testCase := range testCases {
+			testCase := testCase
+
+			t.Run(testCase.name, func(t *testing.T) {
+				t.Parallel()
+
+				event := trace.Event{
+					EventID: int(Ptrace),
+					Args:    testCase.args,
+				}
+				err := ParseArgs(&event)
+				require.NoError(t, err)
+				for _, expArg := range testCase.expectedArgs {
+					arg := GetArg(&event, expArg.Name)
+					assert.Equal(t, expArg, *arg)
+				}
+			})
+		}
+	})
+
 	t.Run("Parse setsockopt value", func(t *testing.T) {
 		t.Parallel()