### 1. Explain what the PR does

8d86c0c5d **chore(events): optimize ParseArgs for uintptr conv**
d4144ae50 **chore(events): improve flag parsing performance**


8d86c0c5d **chore(events): optimize ParseArgs for uintptr conv**

```
- Improved execution time by 20.78% and reduced byte operations by 55%.
- Changed loop to iterate by index for direct modification of
  argument values instead of iterating again via SetArgValue.
- Replaced string concatenation with strconv.AppendUint for efficient
  string construction.

Running tool: /home/gg/.goenv/versions/1.22.4/bin/go test -benchmem
-run=^$ -tags ebpf
-bench ^(BenchmarkParseArgs_Uintptr|BenchmarkParseArgs_Old)$
github.com/aquasecurity/tracee/pkg/events -benchtime=10000000x -race

goos: linux
goarch: amd64
pkg: github.com/aquasecurity/tracee/pkg/events
cpu: AMD Ryzen 9 7950X 16-Core Processor
=== RUN   BenchmarkParseArgs_Uintptr
BenchmarkParseArgs_Uintptr
BenchmarkParseArgs_Uintptr-32     10000000               609.9 ns/op           144 B/op          7 allocs/op
=== RUN   BenchmarkParseArgs_Old
BenchmarkParseArgs_Old
BenchmarkParseArgs_Old-32         10000000               769.7 ns/op           320 B/op          8 allocs/op
PASS
ok      github.com/aquasecurity/tracee/pkg/events       14.829s
```

d4144ae50 **chore(events): improve flag parsing performance**

```
Move helpers functions out of the ParseArgs method which reduces the
overhead of defining closures on each call to ParseArgs.

It reduces ~13,3% of processing time in 10mi iterations, as benchmarked
below:

Running tool: /home/gg/.goenv/versions/1.22.4/bin/go test -benchmem
-run=^$ -tags ebpf -bench ^(BenchmarkParseArgs|BenchmarkParseArgsOld)$
github.com/aquasecurity/tracee/pkg/events -benchtime=10000000x -race

goos: linux
goarch: amd64
pkg: github.com/aquasecurity/tracee/pkg/events
cpu: AMD Ryzen 9 7950X 16-Core Processor
=== RUN   BenchmarkParseArgs
BenchmarkParseArgs
BenchmarkParseArgs-32           10000000 2190 ns/op  0 B/op  0 allocs/op
=== RUN   BenchmarkParseArgsOld
BenchmarkParseArgsOld
BenchmarkParseArgsOld-32        10000000 2526 ns/op  0 B/op  0 allocs/op
PASS
ok      github.com/aquasecurity/tracee/pkg/events       48.189s
```

### 2. Explain how to test it

<!--
Maintainer will review the code, and test the fix/feature, how to run Tracee ?
Give a full command line example and what to look for.
-->

### 3. Other comments

Note that the benchmark measurements are non-deterministic.