@geyslan 2024-07-23
I'm doing other parallel changes but will raise other PR for it, since it'll trigger another performance tests for us.